refactor(roles): change grant_value to cents
changed table roles grant_value to cents
(x100 -> unsigned integer)

+++ database/migrations/2021_06_28_215159_create_roles_table.php
@@ -17,7 +17,7 @@ class CreateRolesTable extends Migration
             $table->id();
             $table->string('name', 50);
             $table->string('description', 110);
-            $table->decimal('grant_value', 8, 2);
+            $table->unsignedInteger('grant_value');
             $table->foreignId('grant_type_id')->constrained('grant_types');
             $table->timestamps();
         });
